requirements.txt
Flask
opencv-python
numpy
pillow
torch
ultralytics
